Revert "Dependabot/GitHub actions/smartcontractkit/push gha metrics action 2 …" (#35)
This reverts commit b19af3cc06d6483da83a8bc06e4ea3a505445dca.

+++ contracts/src/v0.8/shared/interfaces/IWERC20.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.0;
 interface IWERC20 {
   function deposit() external payable;
 
-  function withdraw(uint256) external;
+  function withdraw(uint) external;
 }
